Develop d

+++ src/main/java/com/base/basesetup/entity/ChargerCostInvoiceVO.java
@@ -60,7 +60,7 @@ public class ChargerCostInvoiceVO {
 	@Column(name = "qty")
 	private int qty;
 	@Column(name = "gstpercent")
-	private Float GSTPercent;
+	private double GSTPercent;
 	@Column(name = "tlcamount", precision = 10, scale = 2)
 	private BigDecimal tlcAmount;
 	@Column(name = "gstamount", precision = 10, scale = 2)
